move test for prepareCreateDbInput to CreateWorker tests
