+++ services/rds/rds_test.go
@@ -84,181 +84,21 @@ func NewTestDedicatedDBAdapter(s *config.Settings, rdsClient RDSClientInterface,
 	return NewRdsDedicatedDBAdapter(ctx, s, brokerDB, rdsClient, parameterGroupClient, logger, riverClient)
 }
 
-func TestPrepareCreateDbInstanceInput(t *testing.T) {
-	testErr := errors.New("fail")
-	testCases := map[string]struct {
-		dbInstance        *RDSInstance
-		dbAdapter         *dedicatedDBAdapter
-		expectedGroupName string
-		expectedErr       error
-		password          string
-		expectedParams    *rds.CreateDBInstanceInput
-		plan              *catalog.RDSPlan
-		tags              map[string]string
-	}{
-		"expect error": {
-			dbInstance: &RDSInstance{
-				BinaryLogFormat: "ROW",
-				DbType:          "mysql",
-				dbUtils:         &RDSDatabaseUtils{},
-			},
-			dbAdapter: NewTestDedicatedDBAdapter(
-				&config.Settings{},
-				&mockRDSClient{},
-				&mockParameterGroupClient{
-					returnErr: testErr,
-					rds:       &mockRDSClient{},
-				},
-			),
-			plan:        &catalog.RDSPlan{},
-			expectedErr: testErr,
-		},
-		"creates correct params": {
-			dbInstance: &RDSInstance{
-				AllocatedStorage: 10,
-				Database:         "db-1",
-				BinaryLogFormat:  "ROW",
-				DbType:           "mysql",
-				dbUtils: &MockDbUtils{
-					mockFormattedDbName: "formatted-name",
-				},
-				Username:              "fake-user",
-				StorageType:           "storage-1",
-				PubliclyAccessible:    true,
-				BackupRetentionPeriod: 14,
-				DbSubnetGroup:         "subnet-group-1",
-				SecGroup:              "sec-group-1",
-			},
-			tags: map[string]string{
-				"foo": "bar",
-			},
-			dbAdapter: NewTestDedicatedDBAdapter(
-				&config.Settings{
-					PubliclyAccessibleFeature: true,
-				},
-				&mockRDSClient{},
-				&mockParameterGroupClient{
-					rds:              &mockRDSClient{},
-					customPgroupName: "parameter-group-1",
-				},
-			),
-			plan: &catalog.RDSPlan{
-				InstanceClass: "class-1",
-				Redundant:     true,
-				Encrypted:     true,
-			},
-			password: "fake-password",
-			expectedParams: &rds.CreateDBInstanceInput{
-				AllocatedStorage:        aws.Int32(10),
-				DBInstanceClass:         aws.String("class-1"),
-				DBInstanceIdentifier:    aws.String("db-1"),
-				DBName:                  aws.String("formatted-name"),
-				Engine:                  aws.String("mysql"),
-				MasterUserPassword:      aws.String("fake-password"),
-				MasterUsername:          aws.String("fake-user"),
-				AutoMinorVersionUpgrade: aws.Bool(true),
-				MultiAZ:                 aws.Bool(true),
-				StorageEncrypted:        aws.Bool(true),
-				StorageType:             aws.String("storage-1"),
-				Tags: []rdsTypes.Tag{
-					{
-						Key:   aws.String("foo"),
-						Value: aws.String("bar"),
-					},
-				},
-				PubliclyAccessible:    aws.Bool(true),
-				BackupRetentionPeriod: aws.Int32(14),
-				DBSubnetGroupName:     aws.String("subnet-group-1"),
-				VpcSecurityGroupIds: []string{
-					"sec-group-1",
-				},
-				DBParameterGroupName: aws.String("parameter-group-1"),
-			},
-		},
-		"handles optional params": {
-			dbInstance: &RDSInstance{
-				AllocatedStorage: 10,
-				Database:         "db-1",
-				BinaryLogFormat:  "ROW",
-				DbType:           "mysql",
-				DbVersion:        "8.0",
-				dbUtils: &MockDbUtils{
-					mockFormattedDbName: "formatted-name",
-				},
-				Username:              "fake-user",
-				StorageType:           "storage-1",
-				PubliclyAccessible:    true,
-				BackupRetentionPeriod: 14,
-				DbSubnetGroup:         "subnet-group-1",
-				SecGroup:              "sec-group-1",
-				LicenseModel:          "foo",
-			},
-			tags: map[string]string{
-				"foo": "bar",
-			},
-			dbAdapter: NewTestDedicatedDBAdapter(
-				&config.Settings{
-					PubliclyAccessibleFeature: true,
-				},
-				&mockRDSClient{},
-				&mockParameterGroupClient{
-					rds:              &mockRDSClient{},
-					customPgroupName: "parameter-group-1",
-				},
-			),
-			plan: &catalog.RDSPlan{
-				InstanceClass: "class-1",
-				Redundant:     true,
-				Encrypted:     true,
-			},
-			password: "fake-password",
-			expectedParams: &rds.CreateDBInstanceInput{
-				AllocatedStorage:        aws.Int32(10),
-				DBInstanceClass:         aws.String("class-1"),
-				DBInstanceIdentifier:    aws.String("db-1"),
-				DBName:                  aws.String("formatted-name"),
-				Engine:                  aws.String("mysql"),
-				MasterUserPassword:      aws.String("fake-password"),
-				MasterUsername:          aws.String("fake-user"),
-				AutoMinorVersionUpgrade: aws.Bool(true),
-				MultiAZ:                 aws.Bool(true),
-				StorageEncrypted:        aws.Bool(true),
-				StorageType:             aws.String("storage-1"),
-				Tags: []rdsTypes.Tag{
-					{
-						Key:   aws.String("foo"),
-						Value: aws.String("bar"),
-					},
-				},
-				PubliclyAccessible:    aws.Bool(true),
-				BackupRetentionPeriod: aws.Int32(14),
-				DBSubnetGroupName:     aws.String("subnet-group-1"),
-				VpcSecurityGroupIds: []string{
-					"sec-group-1",
-				},
-				DBParameterGroupName: aws.String("parameter-group-1"),
-				EngineVersion:        aws.String("8.0"),
-				LicenseModel:         aws.String("foo"),
-			},
-		},
-	}
-
-	for name, test := range testCases {
-		t.Run(name, func(t *testing.T) {
-			test.dbInstance.setTags(test.plan, test.tags)
-			params, err := test.dbAdapter.prepareCreateDbInput(test.dbInstance, test.plan, test.password)
-			if err != nil && test.expectedErr == nil {
-				t.Errorf("expected error: %s, got: %s", test.expectedErr, err)
-			}
-			if test.expectedErr != nil && (err == nil || err.Error() != test.expectedErr.Error()) {
-				t.Errorf("expected error: %s, got: %s", test.expectedErr, err)
-			}
-			if diff := deep.Equal(params, test.expectedParams); diff != nil {
-				t.Error(diff)
-			}
-		})
-	}
-}
+// 		t.Run(name, func(t *testing.T) {
+// 			test.dbInstance.setTags(test.plan, test.tags)
+// 			params, err := test.dbAdapter.prepareCreateDbInput(test.dbInstance, test.plan, test.password)
+// 			if err != nil && test.expectedErr == nil {
+// 				t.Errorf("expected error: %s, got: %s", test.expectedErr, err)
+// 			}
+// 			if test.expectedErr != nil && (err == nil || err.Error() != test.expectedErr.Error()) {
+// 				t.Errorf("expected error: %s, got: %s", test.expectedErr, err)
+// 			}
+// 			if diff := deep.Equal(params, test.expectedParams); diff != nil {
+// 				t.Error(diff)
+// 			}
+// 		})
+// 	}
+// }
 
 func TestAsyncCreateDb(t *testing.T) {
 	createDbErr := errors.New("create DB error")
